refactor(provisa-ui): decompose RelationshipsPage into components/relationships
Extract the add form, the expandable detail/edit row, the reverse + conflict
modals, the AI-candidates table, and shared form types out of the 1551-line
RelationshipsPage into components/relationships. Page drops to ~550 lines;
filtering and pagination behavior preserved.
